backend/requirements.txt
feedparser==6.0.11
google-generativeai==0.8.3
requests==2.32.3
firebase-admin==6.6.0
flask==3.1.0
flask-cors==5.0.0
APScheduler==3.10.4
python-dotenv==1.0.1
gunicorn==21.2.0
